pyproject.toml
[project]
name = "medical-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.1",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.8",
    "langgraph>=0.1.5",
    "langsmith>=0.1.75",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "pdfplumber>=0.11.0",
    "pillow>=10.3.0",
    "pinecone>=8.1.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.35.0",
]
